requirements.txt
pyTelegramBotAPI
python-dotenv
pynse